moving components in to folder
