+++ cake/src/client/components/CakeModal.js
