+++ cake/src/client/components/CakeSummary.js
